+++ TongbaoExchangeCalc/Impl/Simulation/ExchangeRecord.cs
@@ -14,11 +14,11 @@ namespace TongbaoExchangeCalc.Impl.Simulation
     public unsafe struct ExchangeResultRecord
     {
         public Int16 TongbaoId; // <10000,2B, value: tongbaoId after exchange
-        public byte SlotIndex; // 0~12,1B
+        public sbyte SlotIndex; // 0~12,1B
         public ExchangeStepResult ExchangeStepResult; // 1B
 
         // 确定数组大小，用的时候不用new
-        public fixed Int16 ResRecords[(byte)ResType.Count - 1]; // index: (byte)ResType-1, value: resValue after exchange
+        public fixed Int16 ResRecords[(int)ResType.Count - 1]; // index: (byte)ResType-1, value: resValue after exchange
     }
 
     public struct ExchangeRecord
